New Layer type called Water
Players cannot cross this.
Because the server is not aware of it, anything else can.
In future Fireballs *should* cross it.
Carrots maybe should not grow in it?
Server dev can decide how to handle it, but at least now
Fireballs cross it and players do not, which makes me happy.