Move spell casting to the Player Object
Now it happens when a player update is sent, with the player data,
such as location, instead of instantly.
NOTE: Currently spell responsese from the server are logged to the
console, to help with server side coding.
The client implementation here is probably borken in its current state,
but this should be enough to help the server side work.
Then I can clean up any client side problems and remove the console
logging.