Background color behavior change in Phaser.
Starting with 3.5, the camera background affects the GAME play area instead of behind it.

Also the method I was using of painting over the screen with one color before adding tiles wasn't working. We got a black background when doing a Refresh on any scene OTHER THAN the default scene.

Now the camera background is set to white, which makes the game area not covered by tiles white, as is required by the Hyrule maps that don't have "ground cover" for the entire map.