Add two things: 1. A loading screen!! 2. A "load testing" feature.
The Loading Screen is a bit perfunctory at the moment. It may actually delay the game loading as it wizzes by so fast, but it performs two functions:
1. It gives a landing spot for players joining so that they don't momentarily slip through the primary game scene. This is because Phaser will ALWAYS load the player into the first scene in the scene array, so putting this at the top does that.
2. IF the game ever does start to get slow to load, this will preemptively make it clear what is going on to users and developers, instead of just waiting at a black screen wondering what up.

The "Load Testing" bit just starts your character firing rapidly and continuously and turns your health to 1000 so that you can start up a bunch of players and spam the server with hadrons.